server/requirements.txt
aiofiles==0.8.0
alembic==1.7.6
black==22.1.0
blinker==1.4
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
dash==2.1.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-extensions==0.0.69
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
EditorConfig==0.12.3
ffmpeg-python==0.2.0
flake8==4.0.1
Flask==2.0.2
Flask-Caching==1.10.1
Flask-Compress==1.10.1
flask-sock==0.4.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.1.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.4
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsbeautifier==1.14.0
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.12.0
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
plotly==5.5.0
priority==2.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.0
pyee==9.0.4
pyflakes==2.4.0
pyparsing==3.0.7
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
simple-websocket==0.5.0
six==1.16.0
SQLAlchemy==1.4.31
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.0.0
zipp==3.7.0
